refactor llm generate service and add polish api
