+++ algorithm/lazymind/chat/api/__init__.py
@@ -1,10 +1,11 @@
 from __future__ import annotations
 
-from . import chat_routes, health_routes, model_check_routes, model_features_routes
+from . import chat_routes, health_routes, llm_generate_routes, model_check_routes, model_features_routes
 
 __all__ = [
     'chat_routes',
     'health_routes',
+    'llm_generate_routes',
     'model_check_routes',
     'model_features_routes',
 ]
