+++ algorithm/lazymind/chat/engine/tools/__init__.py
@@ -6,10 +6,10 @@ registration side effects happen in one consistent place.
 
 from .calculator import calculator
 from .kb import KBToolGroup, TempKBToolGroup
-from .memory import memory_editor
+from .memory_editor import memory_editor
 from .multimodal import vision_extractor
-from .skill_manager import skill_editor
-from .vocab import vocab_learn
+from .skill_editor import skill_editor
+from .vocab_learn import vocab_learn
 from .web_search import url_fetch
 
 __all__ = [
