+++ algorithm/lazymind/chat/engine/tools/skill_editor.py
