+++ algorithm/lazymind/chat/service/__init__.py
@@ -1,3 +1,15 @@
 """Chat service-layer modules."""
 
-__all__: list[str] = []
+from .llm_generate import (
+    BadRequestError,
+    LlmGenerateTaskType,
+    UnprocessableContentError,
+    generate_llm_content,
+)
+
+__all__ = [
+    'BadRequestError',
+    'LlmGenerateTaskType',
+    'UnprocessableContentError',
+    'generate_llm_content',
+]
