+++ algorithm/lazymind/chat/service/llm_generate/__init__.py
@@ -1,0 +1,23 @@
+"""LLM generate service — unified content generation for skill, memory, user_preference, and polish."""
+
+from __future__ import annotations
+
+from .base import (
+    BadRequestError,
+    LlmGenerateTaskType,
+    UnprocessableContentError,
+    generate_llm_content,
+)
+
+# Import business modules to register their prompt builders and edit dispatch
+from . import skill  # noqa: F401
+from . import memory  # noqa: F401
+from . import preference  # noqa: F401
+from . import polish  # noqa: F401
+
+__all__ = [
+    'BadRequestError',
+    'LlmGenerateTaskType',
+    'UnprocessableContentError',
+    'generate_llm_content',
+]
