+++ algorithm/lazymind/review/api/__init__.py
@@ -1,8 +1,7 @@
 from __future__ import annotations
 
-from . import memory_generate_routes, vocab_routes
+from . import vocab_routes
 
 __all__ = [
-    'memory_generate_routes',
     'vocab_routes',
 ]
