--- algorithm/lazymind/review/service/memory_generate/editors/__init__.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-from .memory import _apply_memory_edit_operations
-
-__all__ = [
-    '_apply_memory_edit_operations',
-]
