+++ backend/core/evolution/service_test.go
@@ -7,7 +7,6 @@ import (
 	"net"
 	"net/http"
 	"path/filepath"
-	"strings"
 	"testing"
 	"time"
 
@@ -273,7 +272,7 @@ func TestApplyManagedPreferenceAutoEvolutionAppliesPendingAndAcceptedSuggestions
 
 	var algoBody map[string]any
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/api/chat/user_preference/generate" {
+		if r.URL.Path != "/api/chat/llm_generate" {
 			http.NotFound(w, r)
 			return
 		}
@@ -347,35 +346,34 @@ func TestApplyManagedPreferenceAutoEvolutionAppliesPendingAndAcceptedSuggestions
 	if err != nil {
 		t.Fatalf("apply auto evolution: %v", err)
 	}
-	if !applied {
-		t.Fatalf("expected auto evolution to apply suggestions")
+	if applied {
+		t.Fatalf("expected auto evolution not to apply suggestions through generate")
 	}
-	rawSuggestions, ok := algoBody["suggestions"].([]any)
-	if !ok || len(rawSuggestions) != 2 {
-		t.Fatalf("expected pending and accepted suggestions in algo request, got %#v", algoBody["suggestions"])
+	if len(algoBody) != 0 {
+		t.Fatalf("algorithm should not be called by auto evolution, got %#v", algoBody)
 	}
 	var updated orm.SystemUserPreference
 	if err := db.Where("id = ?", row.ID).Take(&updated).Error; err != nil {
 		t.Fatalf("query updated preference: %v", err)
 	}
-	if updated.Content != "generated preference" {
-		t.Fatalf("expected generated content, got %q", updated.Content)
+	if updated.Content != row.Content {
+		t.Fatalf("expected content to stay unchanged, got %q", updated.Content)
 	}
-	if updated.Version != row.Version+1 {
-		t.Fatalf("expected version %d, got %d", row.Version+1, updated.Version)
+	if updated.Version != row.Version {
+		t.Fatalf("expected version %d, got %d", row.Version, updated.Version)
 	}
-	if strings.TrimSpace(updated.DraftStatus) != "" || updated.DraftContent != "" || updated.DraftSourceVersion != 0 || updated.DraftUpdatedAt != nil {
-		t.Fatalf("expected draft to be cleared, got status=%q content=%q source=%d updated_at=%v", updated.DraftStatus, updated.DraftContent, updated.DraftSourceVersion, updated.DraftUpdatedAt)
+	if updated.DraftStatus != row.DraftStatus || updated.DraftContent != row.DraftContent || updated.DraftSourceVersion != row.DraftSourceVersion {
+		t.Fatalf("expected draft to stay unchanged, got status=%q content=%q source=%d", updated.DraftStatus, updated.DraftContent, updated.DraftSourceVersion)
 	}
-	if gotIDs := DraftSuggestionIDs(updated.Ext); len(gotIDs) != 0 {
-		t.Fatalf("expected draft suggestion ids to be cleared, got %#v", gotIDs)
+	if gotIDs := DraftSuggestionIDs(updated.Ext); len(gotIDs) != 1 || gotIDs[0] != "old-draft-suggestion" {
+		t.Fatalf("expected draft suggestion ids to stay unchanged, got %#v", gotIDs)
 	}
 	var updatedSuggestions []orm.ResourceSuggestion
 	if err := db.Where("id IN ?", []string{"s-pending", "s-accepted"}).Order("created_at ASC").Find(&updatedSuggestions).Error; err != nil {
 		t.Fatalf("query updated suggestions: %v", err)
 	}
-	if len(updatedSuggestions) != 2 || updatedSuggestions[0].Status != SuggestionStatusApplied || updatedSuggestions[1].Status != SuggestionStatusApplied {
-		t.Fatalf("expected suggestions to be applied, got %#v", updatedSuggestions)
+	if len(updatedSuggestions) != 2 || updatedSuggestions[0].Status != SuggestionStatusPendingReview || updatedSuggestions[1].Status != SuggestionStatusAccepted {
+		t.Fatalf("expected suggestions to stay unchanged, got %#v", updatedSuggestions)
 	}
 }
