+++ backend/core/openapi_registry.go
@@ -669,8 +669,7 @@ type listTasksQueryParams struct {
 }
 
 type skillGenerateOpenAPIRequest struct {
-	SuggestionIDs []string `json:"suggestion_ids"`
-	UserInstruct  string   `json:"user_instruct"`
+	UserInstruct string `json:"user_instruct"`
 }
 
 type skillGenerateOpenAPIResponse struct {
@@ -1046,10 +1045,9 @@ type personalizationSettingOpenAPIResponse struct {
 }
 
 type systemGenerateOpenAPIResponse struct {
-	DraftStatus        string   `json:"draft_status"`
-	DraftSourceVersion int64    `json:"draft_source_version"`
-	DraftContent       string   `json:"draft_content"`
-	SuggestionIDs      []string `json:"suggestion_ids"`
+	DraftStatus        string `json:"draft_status"`
+	DraftSourceVersion int64  `json:"draft_source_version"`
+	DraftContent       string `json:"draft_content"`
 }
 
 type systemDraftPreviewOpenAPIResponse struct {
