+++ backend/core/preference/handler.go
@@ -15,6 +15,7 @@ import (
 	"lazymind/core/common/orm"
 	"lazymind/core/evolution"
 	appLog "lazymind/core/log"
+	"lazymind/core/modelconfig"
 	"lazymind/core/store"
 )
 
@@ -24,8 +25,7 @@ type suggestionRequest struct {
 }
 
 type generateRequest struct {
-	SuggestionIDs []string `json:"suggestion_ids"`
-	UserInstruct  string   `json:"user_instruct"`
+	UserInstruct string `json:"user_instruct"`
 }
 
 type upsertRequest struct {
@@ -447,10 +447,9 @@ func Generate(w http.ResponseWriter, r *http.Request) {
 		common.ReplyErr(w, "invalid body", http.StatusBadRequest)
 		return
 	}
-	req.SuggestionIDs = compactIDs(req.SuggestionIDs)
 	req.UserInstruct = strings.TrimSpace(req.UserInstruct)
-	if len(req.SuggestionIDs) == 0 && req.UserInstruct == "" {
-		common.ReplyErr(w, "suggestion_ids or user_instruct required", http.StatusBadRequest)
+	if req.UserInstruct == "" {
+		common.ReplyErr(w, "user_instruct required", http.StatusBadRequest)
 		return
 	}
 
@@ -459,31 +458,22 @@ func Generate(w http.ResponseWriter, r *http.Request) {
 		common.ReplyErr(w, "query user_preference failed", http.StatusInternalServerError)
 		return
 	}
-	useDraft := len(req.SuggestionIDs) == 0 && req.UserInstruct != ""
-	content, err := preferenceGenerateBaseContent(*row, useDraft)
+	content, err := preferenceGenerateBaseContent(*row)
 	if err != nil {
 		common.ReplyErr(w, err.Error(), http.StatusNotFound)
 		return
 	}
 
-	var suggestions []orm.ResourceSuggestion
-	if len(req.SuggestionIDs) > 0 {
-		suggestions, err = evolution.LoadApprovedSuggestions(r.Context(), db, userID, evolution.ResourceTypeUserPreference, evolution.SystemResourceKey(evolution.ResourceTypeUserPreference), req.SuggestionIDs)
-		if err != nil {
-			common.ReplyErr(w, "query suggestions failed", http.StatusInternalServerError)
-			return
-		}
-		if len(suggestions) == 0 {
-			common.ReplyErr(w, "no accepted suggestions found", http.StatusBadRequest)
-			return
-		}
-	}
-
-	algoReq := algo.MemoryGenerateRequest{
+	algoReq := algo.ManagedGenerateRequest{
 		Content:      content,
-		Suggestions:  toAlgoSuggestions(suggestions),
 		UserInstruct: req.UserInstruct,
 	}
+	llmConfig, err := modelconfig.LoadLLMConfig(r.Context(), db, userID)
+	if err != nil {
+		common.ReplyErr(w, "load llm config failed", http.StatusInternalServerError)
+		return
+	}
+	algoReq.LLMConfig = llmConfig
 	appLog.Logger.Info().
 		Str("route", "/user-preference:generate").
 		Str("user_preference_id", row.ID).
@@ -497,10 +487,6 @@ func Generate(w http.ResponseWriter, r *http.Request) {
 	}
 
 	now := time.Now()
-	ids := suggestionIDs(suggestions)
-	if useDraft && len(ids) == 0 {
-		ids = evolution.DraftSuggestionIDs(row.Ext)
-	}
 	update := map[string]any{
 		"draft_content":        generated,
 		"draft_source_version": row.Version,
@@ -509,7 +495,7 @@ func Generate(w http.ResponseWriter, r *http.Request) {
 		"updated_by":           userID,
 		"updated_by_name":      userName,
 		"updated_at":           now,
-		"ext":                  evolution.WithDraftSuggestionIDs(row.Ext, ids),
+		"ext":                  evolution.WithDraftSuggestionIDs(row.Ext, nil),
 	}
 	if err := db.WithContext(r.Context()).Model(&orm.SystemUserPreference{}).Where("id = ?", row.ID).Updates(update).Error; err != nil {
 		common.ReplyErr(w, "update user_preference draft failed", http.StatusInternalServerError)
@@ -519,18 +505,14 @@ func Generate(w http.ResponseWriter, r *http.Request) {
 		"draft_status":         "pending_confirm",
 		"draft_source_version": row.Version,
 		"draft_content":        generated,
-		"suggestion_ids":       ids,
 	})
 }
 
-func preferenceGenerateBaseContent(row orm.SystemUserPreference, useDraft bool) (string, error) {
-	if !useDraft {
-		return row.Content, nil
+func preferenceGenerateBaseContent(row orm.SystemUserPreference) (string, error) {
+	if strings.TrimSpace(row.DraftStatus) == "pending_confirm" {
+		return row.DraftContent, nil
 	}
-	if strings.TrimSpace(row.DraftStatus) != "pending_confirm" {
-		return "", errors.New("user_preference draft not found")
-	}
-	return row.DraftContent, nil
+	return row.Content, nil
 }
 
 func Confirm(w http.ResponseWriter, r *http.Request) {
@@ -560,7 +542,6 @@ func Confirm(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	ids := evolution.DraftSuggestionIDs(row.Ext)
 	now := time.Now()
 	newContent := row.DraftContent
 	update := map[string]any{
@@ -578,10 +559,6 @@ func Confirm(w http.ResponseWriter, r *http.Request) {
 	}
 	if err := db.WithContext(r.Context()).Model(&orm.SystemUserPreference{}).Where("id = ? AND version = ?", row.ID, row.Version).Updates(update).Error; err != nil {
 		common.ReplyErr(w, "confirm user_preference draft failed", http.StatusInternalServerError)
-		return
-	}
-	if err := evolution.UpdateSuggestionStatus(r.Context(), db, ids, evolution.SuggestionStatusApplied); err != nil {
-		common.ReplyErr(w, "update suggestion status failed", http.StatusInternalServerError)
 		return
 	}
 	common.ReplyOK(w, map[string]any{
@@ -638,43 +615,4 @@ func firstNonEmpty(values ...string) string {
 		}
 	}
 	return ""
-}
-
-func compactIDs(ids []string) []string {
-	seen := make(map[string]struct{}, len(ids))
-	out := make([]string, 0, len(ids))
-	for _, id := range ids {
-		id = strings.TrimSpace(id)
-		if id == "" {
-			continue
-		}
-		if _, ok := seen[id]; ok {
-			continue
-		}
-		seen[id] = struct{}{}
-		out = append(out, id)
-	}
-	return out
-}
-
-func suggestionIDs(rows []orm.ResourceSuggestion) []string {
-	out := make([]string, 0, len(rows))
-	for _, row := range rows {
-		if strings.TrimSpace(row.ID) != "" {
-			out = append(out, strings.TrimSpace(row.ID))
-		}
-	}
-	return out
-}
-
-func toAlgoSuggestions(rows []orm.ResourceSuggestion) []algo.Suggestion {
-	out := make([]algo.Suggestion, 0, len(rows))
-	for _, row := range rows {
-		out = append(out, algo.Suggestion{
-			Title:   row.Title,
-			Content: row.Content,
-			Reason:  row.Reason,
-		})
-	}
-	return out
 }
