+++ backend/core/preference/handler_test.go
@@ -362,7 +362,7 @@ func TestGenerateOverwritesExistingPendingDraft(t *testing.T) {
 	t.Cleanup(func() { store.Init(nil, nil, nil) })
 
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/api/chat/user_preference/generate" {
+		if r.URL.Path != "/api/chat/llm_generate" {
 			http.NotFound(w, r)
 			return
 		}
@@ -416,7 +416,7 @@ func TestGenerateOverwritesExistingPendingDraft(t *testing.T) {
 		t.Fatalf("create suggestion: %v", err)
 	}
 
-	req := httptest.NewRequest(http.MethodPost, "/api/core/user-preference:generate", strings.NewReader(`{"suggestion_ids":["suggestion-1"],"user_instruct":"生成新版"}`))
+	req := httptest.NewRequest(http.MethodPost, "/api/core/user-preference:generate", strings.NewReader(`{"user_instruct":"生成新版"}`))
 	req.Header.Set("Content-Type", "application/json")
 	req.Header.Set("X-User-Id", "u1")
 	req.Header.Set("X-User-Name", "User 1")
@@ -452,19 +452,19 @@ func TestGenerateOverwritesExistingPendingDraft(t *testing.T) {
 		t.Fatalf("expected draft source version %d, got %d", row.Version, updated.DraftSourceVersion)
 	}
 	gotIDs := evolution.DraftSuggestionIDs(updated.Ext)
-	if len(gotIDs) != 1 || gotIDs[0] != "suggestion-1" {
-		t.Fatalf("expected draft suggestion ids to be replaced, got %#v", gotIDs)
+	if len(gotIDs) != 0 {
+		t.Fatalf("expected draft suggestion ids to be cleared, got %#v", gotIDs)
 	}
 }
 
-func TestGenerateAllowsSuggestionsWithoutUserInstruct(t *testing.T) {
+func TestGenerateRejectsMissingUserInstruct(t *testing.T) {
 	db := newPreferenceTestDB(t)
 	store.Init(db.DB, nil, nil)
 	t.Cleanup(func() { store.Init(nil, nil, nil) })
 
 	var algoBody map[string]any
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/api/chat/user_preference/generate" {
+		if r.URL.Path != "/api/chat/llm_generate" {
 			http.NotFound(w, r)
 			return
 		}
@@ -517,7 +517,7 @@ func TestGenerateAllowsSuggestionsWithoutUserInstruct(t *testing.T) {
 		t.Fatalf("create suggestion: %v", err)
 	}
 
-	req := httptest.NewRequest(http.MethodPost, "/api/core/user-preference:generate", strings.NewReader(`{"suggestion_ids":["suggestion-1"],"user_instruct":""}`))
+	req := httptest.NewRequest(http.MethodPost, "/api/core/user-preference:generate", strings.NewReader(`{"user_instruct":""}`))
 	req.Header.Set("Content-Type", "application/json")
 	req.Header.Set("X-User-Id", "u1")
 	req.Header.Set("X-User-Name", "User 1")
@@ -525,15 +525,11 @@ func TestGenerateAllowsSuggestionsWithoutUserInstruct(t *testing.T) {
 
 	Generate(rec, req)
 
-	if rec.Code != http.StatusOK {
-		t.Fatalf("expected status 200, got %d body=%s", rec.Code, rec.Body.String())
+	if rec.Code != http.StatusBadRequest {
+		t.Fatalf("expected status 400, got %d body=%s", rec.Code, rec.Body.String())
 	}
-	if algoBody["user_instruct"] != "" {
-		t.Fatalf("expected empty user_instruct sent to algorithm, got %#v", algoBody["user_instruct"])
-	}
-	suggestions, ok := algoBody["suggestions"].([]any)
-	if !ok || len(suggestions) != 1 {
-		t.Fatalf("expected one suggestion sent to algorithm, got %#v", algoBody["suggestions"])
+	if len(algoBody) != 0 {
+		t.Fatalf("algorithm should not be called for suggestion-only generate, got %#v", algoBody)
 	}
 }
 
@@ -544,7 +540,7 @@ func TestGenerateUserInstructOnlyUsesDraftContent(t *testing.T) {
 
 	var algoBody map[string]any
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/api/chat/user_preference/generate" {
+		if r.URL.Path != "/api/chat/llm_generate" {
 			http.NotFound(w, r)
 			return
 		}
@@ -618,17 +614,19 @@ func TestGenerateUserInstructOnlyUsesDraftContent(t *testing.T) {
 	if algoBody["content"] != "draft preference" {
 		t.Fatalf("expected draft content sent to algorithm, got %#v", algoBody["content"])
 	}
-	suggestions, ok := algoBody["suggestions"].([]any)
-	if !ok || len(suggestions) != 0 {
-		t.Fatalf("expected empty suggestions array, got %#v", algoBody["suggestions"])
+	if algoBody["task_type"] != "user_preference" {
+		t.Fatalf("expected user_preference task_type, got %#v", algoBody["task_type"])
+	}
+	if _, ok := algoBody["suggestions"]; ok {
+		t.Fatalf("suggestions should not be sent to algorithm: %#v", algoBody["suggestions"])
 	}
 	var updated orm.SystemUserPreference
 	if err := db.Where("id = ?", row.ID).Take(&updated).Error; err != nil {
 		t.Fatalf("query updated preference: %v", err)
 	}
 	gotIDs := evolution.DraftSuggestionIDs(updated.Ext)
-	if len(gotIDs) != 1 || gotIDs[0] != "suggestion-1" {
-		t.Fatalf("expected draft suggestion ids to be preserved, got %#v", gotIDs)
+	if len(gotIDs) != 0 {
+		t.Fatalf("expected draft suggestion ids to be cleared, got %#v", gotIDs)
 	}
 
 	confirmReq := httptest.NewRequest(http.MethodPost, "/api/core/user-preference:confirm", nil)
@@ -645,8 +643,8 @@ func TestGenerateUserInstructOnlyUsesDraftContent(t *testing.T) {
 	if err := db.Where("id = ?", "suggestion-1").Take(&applied).Error; err != nil {
 		t.Fatalf("query applied suggestion: %v", err)
 	}
-	if applied.Status != evolution.SuggestionStatusApplied {
-		t.Fatalf("expected suggestion to be applied after confirm, got %q", applied.Status)
+	if applied.Status != evolution.SuggestionStatusAccepted {
+		t.Fatalf("expected suggestion status to stay accepted after confirm, got %q", applied.Status)
 	}
 }
