+++ backend/core/skill/management_test.go
@@ -469,7 +469,7 @@ func TestGenerateReturnsOutdatedWhenApprovedSuggestionSnapshotIsStale(t *testing
 	t.Cleanup(func() { store.Init(nil, nil, nil) })
 
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/api/chat/skill/generate" {
+		if r.URL.Path != "/api/chat/llm_generate" {
 			http.NotFound(w, r)
 			return
 		}
@@ -548,7 +548,7 @@ func TestGenerateReturnsOutdatedWhenApprovedSuggestionSnapshotIsStale(t *testing
 		t.Fatalf("create suggestion: %v", err)
 	}
 
-	req := httptest.NewRequest(http.MethodPost, "/api/core/skills/skill-1:generate", strings.NewReader(`{"suggestion_ids":["suggestion-1"],"user_instruct":"请生成新版"}`))
+	req := httptest.NewRequest(http.MethodPost, "/api/core/skills/skill-1:generate", strings.NewReader(`{"user_instruct":"请生成新版"}`))
 	req = mux.SetURLVars(req, map[string]string{"skill_id": "skill-1"})
 	req.Header.Set("Content-Type", "application/json")
 	req.Header.Set("X-User-Id", "u1")
@@ -570,8 +570,8 @@ func TestGenerateReturnsOutdatedWhenApprovedSuggestionSnapshotIsStale(t *testing
 	if resp.Data.DraftStatus != "pending_confirm" {
 		t.Fatalf("expected pending_confirm draft status, got %q", resp.Data.DraftStatus)
 	}
-	if !resp.Data.Outdated {
-		t.Fatalf("expected outdated=true")
+	if resp.Data.Outdated {
+		t.Fatalf("expected outdated=false when generate no longer consumes suggestions")
 	}
 	var updatedSkill orm.SkillResource
 	if err := db.Where("id = ?", "skill-1").Take(&updatedSkill).Error; err != nil {
@@ -740,7 +740,7 @@ func TestGenerateAllowsUserInstructWithoutSuggestions(t *testing.T) {
 
 	var algoBody map[string]any
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/api/chat/skill/generate" {
+		if r.URL.Path != "/api/chat/llm_generate" {
 			http.NotFound(w, r)
 			return
 		}
@@ -818,7 +818,7 @@ func TestGenerateAllowsUserInstructWithoutSuggestions(t *testing.T) {
 		t.Fatalf("create suggestion: %v", err)
 	}
 
-	req := httptest.NewRequest(http.MethodPost, "/api/core/skills/skill-1:generate", strings.NewReader(`{"suggestion_ids":[],"user_instruct":"只按用户意见生成"}`))
+	req := httptest.NewRequest(http.MethodPost, "/api/core/skills/skill-1:generate", strings.NewReader(`{"user_instruct":"只按用户意见生成"}`))
 	req = mux.SetURLVars(req, map[string]string{"skill_id": "skill-1"})
 	req.Header.Set("Content-Type", "application/json")
 	req.Header.Set("X-User-Id", "u1")
@@ -841,9 +841,11 @@ func TestGenerateAllowsUserInstructWithoutSuggestions(t *testing.T) {
 	if _, ok := algoBody["skill_name"]; ok {
 		t.Fatalf("skill_name should not be sent to algorithm: %#v", algoBody["skill_name"])
 	}
-	suggestions, ok := algoBody["suggestions"].([]any)
-	if !ok || len(suggestions) != 0 {
-		t.Fatalf("expected empty suggestions array, got %#v", algoBody["suggestions"])
+	if algoBody["task_type"] != "skill" {
+		t.Fatalf("expected skill task_type, got %#v", algoBody["task_type"])
+	}
+	if _, ok := algoBody["suggestions"]; ok {
+		t.Fatalf("suggestions should not be sent to algorithm: %#v", algoBody["suggestions"])
 	}
 	var updatedSkill orm.SkillResource
 	if err := db.Where("id = ?", "skill-1").Take(&updatedSkill).Error; err != nil {
@@ -856,8 +858,8 @@ func TestGenerateAllowsUserInstructWithoutSuggestions(t *testing.T) {
 		t.Fatalf("expected update_status to stay up_to_date for instruction-only draft, got %q", updatedSkill.UpdateStatus)
 	}
 	gotIDs := evolution.DraftSuggestionIDs(updatedSkill.Ext)
-	if len(gotIDs) != 1 || gotIDs[0] != "suggestion-1" {
-		t.Fatalf("expected draft suggestion ids to be preserved, got %#v", gotIDs)
+	if len(gotIDs) != 0 {
+		t.Fatalf("expected draft suggestion ids to be cleared, got %#v", gotIDs)
 	}
 
 	confirmReq := httptest.NewRequest(http.MethodPost, "/api/core/skills/skill-1:confirm", nil)
@@ -874,8 +876,8 @@ func TestGenerateAllowsUserInstructWithoutSuggestions(t *testing.T) {
 	if err := db.Where("id = ?", "suggestion-1").Take(&applied).Error; err != nil {
 		t.Fatalf("query applied suggestion: %v", err)
 	}
-	if applied.Status != evolution.SuggestionStatusApplied {
-		t.Fatalf("expected suggestion to be applied after confirm, got %q", applied.Status)
+	if applied.Status != evolution.SuggestionStatusAccepted {
+		t.Fatalf("expected suggestion status to stay accepted after confirm, got %q", applied.Status)
 	}
 }
 
@@ -966,7 +968,7 @@ func TestGenerateAllowsGeneratedDescriptionChange(t *testing.T) {
 	t.Cleanup(func() { store.Init(nil, nil, nil) })
 
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/api/chat/skill/generate" {
+		if r.URL.Path != "/api/chat/llm_generate" {
 			http.NotFound(w, r)
 			return
 		}
@@ -1209,8 +1211,8 @@ func TestDraftPreviewReturnsCurrentDraftAndDiff(t *testing.T) {
 	if !strings.Contains(resp.Data.Diff, "+updated body") {
 		t.Fatalf("expected diff to contain added draft line, got %q", resp.Data.Diff)
 	}
-	if !resp.Data.Outdated {
-		t.Fatalf("expected outdated=true")
+	if resp.Data.Outdated {
+		t.Fatalf("expected outdated=false")
 	}
 }
